make all main content go in a main block
you should be able to switch the {% extends %} and still have a page
render, which means you have to use the same block names everywhere.
Possible block names to use are body, content, and main, in that scope.
Then add in _base blocks and baby, you got a stew going.