do a beauty pass on the pricetrends chart
the markup is kinda weird, but the look is there.

weirdness: having to put the dollar sign in its own span, that's a lot
of markup for very little return. I tried doing it using the :before
pseudoelement but got nowhere.

needing to set margin-right to control whitespace. using margins for
placement is a bad idea.