remove caching sentences on the model
fixes: sentence cache couldn't get evicted
there isn't a template that does a list view of sentences anymore so we
can take the performance hit for now.
I would like to see it come back but use the django cache backend and
have a way to clear it.
Also, even though it was only a few kilobytes, making the web heads hold
onto that information seemed odd to me.